+++ lib/src/ui/columns/trina_column_title.dart
@@ -16,8 +16,8 @@ class TrinaColumnTitle extends TrinaStatefulWidget {
     required this.stateManager,
     required this.column,
     double? height,
-  }) : height = height ?? stateManager.columnHeight,
-       super(key: ValueKey('column_title_${column.key}'));
+  })  : height = height ?? stateManager.columnHeight,
+        super(key: ValueKey('column_title_${column.key}'));
 
   @override
   TrinaColumnTitleState createState() => TrinaColumnTitleState();
@@ -141,10 +141,9 @@ class TrinaColumnTitleState extends TrinaStateWithChange<TrinaColumnTitle> {
           icon: TrinaGridColumnIcon(
             sort: _sort,
             color: style.iconColor,
-            icon:
-                widget.column.enableContextMenu
-                    ? style.columnContextIcon
-                    : style.columnResizeIcon,
+            icon: widget.column.enableContextMenu
+                ? style.columnContextIcon
+                : style.columnResizeIcon,
             ascendingIcon: style.columnAscendingIcon,
             descendingIcon: style.columnDescendingIcon,
           ),
@@ -163,10 +162,10 @@ class TrinaColumnTitleState extends TrinaStateWithChange<TrinaColumnTitle> {
       // Ensure dragging functionality works with custom renderer
       return widget.column.enableColumnDrag
           ? _DraggableWidget(
-            stateManager: stateManager,
-            column: widget.column,
-            child: customTitleWidget,
-          )
+              stateManager: stateManager,
+              column: widget.column,
+              child: customTitleWidget,
+            )
           : customTitleWidget;
     }
 
@@ -175,28 +174,26 @@ class TrinaColumnTitleState extends TrinaStateWithChange<TrinaColumnTitle> {
         Positioned(
           left: 0,
           right: 0,
-          child:
-              widget.column.enableColumnDrag
-                  ? _DraggableWidget(
-                    stateManager: stateManager,
-                    column: widget.column,
-                    child: columnWidget,
-                  )
-                  : columnWidget,
+          child: widget.column.enableColumnDrag
+              ? _DraggableWidget(
+                  stateManager: stateManager,
+                  column: widget.column,
+                  child: columnWidget,
+                )
+              : columnWidget,
         ),
         if (showContextIcon)
           Positioned.directional(
             textDirection: stateManager.textDirection,
             end: -3,
-            child:
-                enableGesture
-                    ? Listener(
-                      onPointerDown: _handleOnPointDown,
-                      onPointerMove: _handleOnPointMove,
-                      onPointerUp: _handleOnPointUp,
-                      child: contextMenuIcon,
-                    )
-                    : contextMenuIcon,
+            child: enableGesture
+                ? Listener(
+                    onPointerDown: _handleOnPointDown,
+                    onPointerMove: _handleOnPointMove,
+                    onPointerUp: _handleOnPointUp,
+                    child: contextMenuIcon,
+                  )
+                : contextMenuIcon,
           ),
       ],
     );
@@ -212,15 +209,14 @@ class TrinaColumnTitleState extends TrinaStateWithChange<TrinaColumnTitle> {
       stateManager: stateManager,
       height: widget.height,
       showContextIcon: showContextIcon,
-      contextMenuIcon:
-          enableGesture
-              ? Listener(
-                onPointerDown: _handleOnPointDown,
-                onPointerMove: _handleOnPointMove,
-                onPointerUp: _handleOnPointUp,
-                child: contextMenuIcon,
-              )
-              : contextMenuIcon,
+      contextMenuIcon: enableGesture
+          ? Listener(
+              onPointerDown: _handleOnPointDown,
+              onPointerMove: _handleOnPointMove,
+              onPointerUp: _handleOnPointUp,
+              child: contextMenuIcon,
+            )
+          : contextMenuIcon,
       isFiltered: isFiltered,
       showContextMenu:
           mounted && widget.column.enableContextMenu ? _showContextMenu : null,
@@ -254,9 +250,9 @@ class TrinaGridColumnIcon extends StatelessWidget {
       case TrinaColumnSort.ascending:
         return ascendingIcon == null
             ? Transform.rotate(
-              angle: 90 * pi / 90,
-              child: const Icon(Icons.sort, color: Colors.green),
-            )
+                angle: 90 * pi / 90,
+                child: const Icon(Icons.sort, color: Colors.green),
+              )
             : ascendingIcon!;
       case TrinaColumnSort.descending:
         return descendingIcon == null
@@ -352,13 +348,13 @@ class _SortableWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return column.enableSorting
         ? MouseRegion(
-          cursor: SystemMouseCursors.click,
-          child: GestureDetector(
-            key: const ValueKey('ColumnTitleSortableGesture'),
-            onTap: _onTap,
-            child: child,
-          ),
-        )
+            cursor: SystemMouseCursors.click,
+            child: GestureDetector(
+              key: const ValueKey('ColumnTitleSortableGesture'),
+              onTap: _onTap,
+              child: child,
+            ),
+          )
         : child;
   }
 }
@@ -413,17 +409,15 @@ class _ColumnWidget extends StatelessWidget {
           child: DecoratedBox(
             decoration: BoxDecoration(
               gradient: column.backgroundGradient, //
-              color:
-                  column.backgroundGradient == null
-                      ? (noDragTarget
-                          ? column.backgroundColor
-                          : style.dragTargetColumnColor)
-                      : null,
+              color: column.backgroundGradient == null
+                  ? (noDragTarget
+                      ? column.backgroundColor
+                      : style.dragTargetColumnColor)
+                  : null,
               border: BorderDirectional(
-                end:
-                    style.enableColumnBorderVertical
-                        ? BorderSide(color: style.borderColor, width: 1.0)
-                        : BorderSide.none,
+                end: style.enableColumnBorderVertical
+                    ? BorderSide(color: style.borderColor, width: 1.0)
+                    : BorderSide.none,
               ),
             ),
             child: Padding(
@@ -566,23 +560,25 @@ class _ColumnTextWidgetState extends TrinaStateWithChange<_ColumnTextWidget> {
       widget.column.titleSpan == null ? widget.column.title : null;
 
   List<InlineSpan> get _children => [
-    if (widget.column.titleSpan != null) widget.column.titleSpan!,
-    if (_isFilteredList)
-      WidgetSpan(
-        alignment: PlaceholderAlignment.middle,
-        child: IconButton(
-          icon: Icon(
-            Icons.filter_alt_outlined,
-            color: stateManager.configuration.style.iconColor,
-            size: stateManager.configuration.style.iconSize,
+        if (widget.column.titleSpan != null) widget.column.titleSpan!,
+        if (_isFilteredList &&
+            stateManager.configuration.style.filterIcon != null)
+          WidgetSpan(
+            alignment: PlaceholderAlignment.middle,
+            child: IconButton(
+              icon: Icon(
+                stateManager.configuration.style.filterIcon!.icon,
+                color: stateManager.configuration.style.iconColor,
+                size: stateManager.configuration.style.iconSize,
+              ),
+              onPressed: _handleOnPressedFilter,
+              constraints: BoxConstraints(
+                maxHeight:
+                    widget.height + (TrinaGridSettings.rowBorderWidth * 2),
+              ),
+            ),
           ),
-          onPressed: _handleOnPressedFilter,
-          constraints: BoxConstraints(
-            maxHeight: widget.height + (TrinaGridSettings.rowBorderWidth * 2),
-          ),
-        ),
-      ),
-  ];
+      ];
 
   @override
   Widget build(BuildContext context) {