+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add filterEnterKeyAction to TrinaColumn for controlling keyboard navigation in column filters ( by @doonfrs )
 * Update row color example ( by @doonfrs )
 * Fix horizontal scrollbar calculation issue for the header & footer ( by @doonfrs )
+* Added filterIcon option to TrinaGridStyleConfig ( by @doonfrs )
 
 ## [1.6.6] - 2025. 03. 22